+++ src/main/java/com/braintreegateway/AmexExpressCheckoutCard.java
@@ -5,6 +5,7 @@ import java.util.Calendar;
 import java.util.List;
 import java.util.ArrayList;
 
+@Deprecated
 public class AmexExpressCheckoutCard implements PaymentMethod {
     private String cardType;
     private String token;