+++ src/main/java/com/braintreegateway/Customer.java
@@ -162,6 +162,7 @@ public class Customer {
         return Collections.unmodifiableList(androidPayCards);
     }
 
+    @Deprecated
     public List<AmexExpressCheckoutCard> getAmexExpressCheckoutCards() {
         return Collections.unmodifiableList(amexExpressCheckoutCards);
     }
@@ -174,6 +175,7 @@ public class Customer {
         return Collections.unmodifiableList(visaCheckoutCards);
     }
 
+    @Deprecated
     public List<MasterpassCard> getMasterpassCards() {
         return Collections.unmodifiableList(masterpassCards);
     }